Refactoring fetchTransactions to context
