+++ App.tsx
@@ -1,6 +1,6 @@
 import 'expo-dev-client';
 import React from 'react';
-import { ThemeProvider, AmplitudeProvider, TransactionsProvider } from '@contexts';
+import { ThemeProvider, AmplitudeProvider } from '@contexts';
 import Routes from '@routes';
 import AppLoading from 'expo-app-loading';
 import Logger from '@utils/logger';
@@ -14,13 +14,11 @@ const App = () => {
 	if (!coinList || !fontsLoaded || walletState.promised) return <AppLoading />;
 
 	return (
-		<TransactionsProvider>
-			<AmplitudeProvider>
-				<ThemeProvider>
-					<Routes />
-				</ThemeProvider>
-			</AmplitudeProvider>
-		</TransactionsProvider>
+		<AmplitudeProvider>
+			<ThemeProvider>
+				<Routes />
+			</ThemeProvider>
+		</AmplitudeProvider>
 	);
 };
